requirements.txt
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
minio==7.2.3
openpyxl==3.1.2